move fp8 code
